修正敌人死亡判定小bug，简化帧率控制实现

+++ src/cn/edu/scau/cmi/glz/VisHelper.java
@@ -147,7 +147,7 @@ public class VisHelper {
     }
 
     // 延迟t毫秒
-    public static void pause(int t) {
+    public static void pause(long t) {
         try {
             Thread.sleep(t);
         }
